v1 - "Bookmark Up" has been mounted for public display
Bookmark Up is a simple app designed for reorganising Chromium-based browser bookmarks. The Browser I use is Brave Browser. Existing Extensions and apps that do similar things seem to NEED access to personal and private data. That seems wrong to me, hence development. Feel free to play with it as you wish, and keep me in the loop, if you like?